# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentForge",
    "description": "AgentForge - AI 驱动的 Fiverr 运营自动化智能助理系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/backup/status": {
      "get": {
        "tags": [
          "backup"
        ],
        "summary": "Get Backup Status",
        "description": "Get backup scheduler status",
        "operationId": "get_backup_status_api_backup_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/backup/create": {
      "post": {
        "tags": [
          "backup"
        ],
        "summary": "Create Backup",
        "description": "Create a new backup",
        "operationId": "create_backup_api_backup_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/backup/restore": {
      "post": {
        "tags": [
          "backup"
        ],
        "summary": "Restore Backup",
        "description": "Restore from a backup",
        "operationId": "restore_backup_api_backup_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/backup/list": {
      "get": {
        "tags": [
          "backup"
        ],
        "summary": "List Backups",
        "description": "List available backups",
        "operationId": "list_backups_api_backup_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backup/{backup_id}": {
      "get": {
        "tags": [
          "backup"
        ],
        "summary": "Get Backup",
        "description": "Get backup details",
        "operationId": "get_backup_api_backup__backup_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "backup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backup Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "backup"
        ],
        "summary": "Delete Backup",
        "description": "Delete a backup",
        "operationId": "delete_backup_api_backup__backup_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "backup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backup Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backup/{backup_id}/verify": {
      "get": {
        "tags": [
          "backup"
        ],
        "summary": "Verify Backup",
        "description": "Verify backup integrity",
        "operationId": "verify_backup_api_backup__backup_id__verify_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "backup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backup Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backup/retention/apply": {
      "post": {
        "tags": [
          "backup"
        ],
        "summary": "Apply Retention Policy",
        "description": "Apply retention policy",
        "operationId": "apply_retention_policy_api_backup_retention_apply_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/backup/stats/summary": {
      "get": {
        "tags": [
          "backup"
        ],
        "summary": "Get Backup Stats",
        "description": "Get backup statistics",
        "operationId": "get_backup_stats_api_backup_stats_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/backup/schedule": {
      "put": {
        "tags": [
          "backup"
        ],
        "summary": "Update Schedule",
        "description": "Update backup schedule configuration",
        "operationId": "update_schedule_api_backup_schedule_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BackupRequest": {
        "properties": {
          "backup_type": {
            "type": "string",
            "title": "Backup Type",
            "default": "full"
          },
          "include_postgres": {
            "type": "boolean",
            "title": "Include Postgres",
            "default": true
          },
          "include_redis": {
            "type": "boolean",
            "title": "Include Redis",
            "default": true
          },
          "include_qdrant": {
            "type": "boolean",
            "title": "Include Qdrant",
            "default": true
          }
        },
        "type": "object",
        "title": "BackupRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RestoreRequest": {
        "properties": {
          "backup_id": {
            "type": "string",
            "title": "Backup Id"
          },
          "include_postgres": {
            "type": "boolean",
            "title": "Include Postgres",
            "default": true
          },
          "include_redis": {
            "type": "boolean",
            "title": "Include Redis",
            "default": true
          },
          "include_qdrant": {
            "type": "boolean",
            "title": "Include Qdrant",
            "default": true
          },
          "drop_existing": {
            "type": "boolean",
            "title": "Drop Existing",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "backup_id"
        ],
        "title": "RestoreRequest"
      },
...
      "ScheduleUpdateRequest": {
        "properties": {
          "daily_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Enabled"
          },
          "daily_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Time"
          },
          "weekly_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Enabled"
          },
          "weekly_day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Day"
          },
          "retention_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention Days"
          }
        },
        "type": "object",
        "title": "ScheduleUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}